fix: Remove logout from redux, add user resource
