--- querybook/webapp/const/userRoles.ts
@@ -1,4 +1,0 @@
-// Keep it the same as const/user_roles.py
-export enum UserRoleType {
-    ADMIN = 0,
-}
